requirementes.txt
dj-database-url==0.4.2
Django==2.0.2
django-admin-rangefilter==0.3.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-dateutil==2.6.1
python-decouple==3.1
pytz==2018.3
six==1.11.0
uwsgitop==0.10
django-crontab==0.7.1
uWSGI==2.0.14
